Update Workflow.Rmd
I manually added code used for preprocessing of DWI-HARDI images, extraction of NODDI metrics, and simulations of the electric fields induced by TMS. The relative paths of the rest of the code does not allow me to build with workflowr. Please let me know if you can rebuild using the final version of this file. I will direct message you to talk about it.